backend/ttsjob/requirements.txt
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3
selenium>=4.11.2
python-dotenv>=1.0.0
urllib3>=2.0.4
click>=8.1.8
aiohttp>=3.11.11
extruct>=0.18.0
python-dateutil>=2.9.0
aiolimiter>=1.2.1
pyminiscraper>=0.1.1
asyncclick>=8.1.8.0
psycopg2-binary>=2.9.10
asyncpg>=0.30.0
google-cloud-logging>=3.11.3
sqlalchemy>=2.0.37
pgvector>=0.3.6
